assistant-service/requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.1
httpx==0.27.2
sse-starlette==2.1.3
python-dotenv==1.0.1
PyJWT==2.8.0
redis[hiredis]==5.0.1

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg>=0.30.0

# AI Provider SDKs
openai==1.54.4
anthropic>=0.40.0
google-generativeai==0.8.3
ollama==0.4.2

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
respx==0.22.0

# Linting
black==24.10.0
flake8==7.1.1
isort==5.13.2
